+++ test/test/datascript.cljs
@@ -414,8 +414,53 @@
       (is (= (d/q '[:find ?n
                     :where [_ :children ?e]
                            [?e :name ?n]] db)
-             #{["Petr"] ["Evgeny"]})))))
+             #{["Petr"] ["Evgeny"]})))
+    
+    (is (thrown-with-msg? js/Error #"Bad attribute :_parent"
+      (d/db-with db0 [{:name "Sergey" :_parent 1}])))))
 
+(deftest test-explode-nested-maps
+  (let [schema { :profile { :db/valueType :db.type/ref }}
+        db     (d/empty-db schema)]
+    (are [tx res] (= (d/q '[:find ?e ?a ?v
+                            :where [?e ?a ?v]]
+                          (d/db-with db tx)) res)
+      [ {:db/id 5 :name "Ivan" :profile {:db/id 7 :email "@2"}} ]
+      #{ [5 :name "Ivan"] [5 :profile 7] [7 :email "@2"] }
+         
+      [ {:name "Ivan" :profile {:email "@2"}} ]
+      #{ [1 :name "Ivan"] [1 :profile 2] [2 :email "@2"] }
+         
+      [ {:email "@2" :_profile {:name "Ivan"}} ]
+      #{ [1 :email "@2"] [2 :name "Ivan"] [2 :profile 1] }
+    ))
+  
+  (testing "multi-valued"
+    (let [schema { :profile { :db/valueType :db.type/ref
+                              :db/cardinality :db.cardinality/many }}
+          db     (d/empty-db schema)]
+      (are [tx res] (= (d/q '[:find ?e ?a ?v
+                              :where [?e ?a ?v]]
+                            (d/db-with db tx)) res)
+        [ {:db/id 5 :name "Ivan" :profile {:db/id 7 :email "@2"}} ]
+        #{ [5 :name "Ivan"] [5 :profile 7] [7 :email "@2"] }
+           
+        [ {:db/id 5 :name "Ivan" :profile [{:db/id 7 :email "@2"} {:db/id 8 :email "@3"}]} ]
+        #{ [5 :name "Ivan"] [5 :profile 7] [7 :email "@2"] [5 :profile 8] [8 :email "@3"] }
+
+        [ {:name "Ivan" :profile {:email "@2"}} ]
+        #{ [1 :name "Ivan"] [1 :profile 2] [2 :email "@2"] }
+
+        [ {:name "Ivan" :profile [{:email "@2"} {:email "@3"}]} ]
+        #{ [1 :name "Ivan"] [1 :profile 2] [2 :email "@2"] [1 :profile 3] [3 :email "@3"] }
+           
+        [ {:email "@2" :_profile {:name "Ivan"}} ]
+        #{ [1 :email "@2"] [2 :name "Ivan"] [2 :profile 1] }
+
+        [ {:email "@2" :_profile [{:name "Ivan"} {:name "Petr"} ]} ]
+        #{ [1 :email "@2"] [2 :name "Ivan"] [2 :profile 1] [3 :name "Petr"] [3 :profile 1] }
+      ))))
+    
 (deftest test-joins
   (let [db (-> (d/empty-db)
                (d/db-with [ { :db/id 1, :name  "Ivan", :age   15 }